Moved a bunch of duplicate functions to a helper class. Fixed oversight in the correct sheet to select for saved schedule.
