+++ workdayspreadsheettoics/src/main/java/workdayspreadsheettoics/SavedScheduleToCalendar.java
@@ -45,42 +45,9 @@ import net.fortuna.ical4j.util.UidGenerator;
 
 
 
+
 public class SavedScheduleToCalendar {
   private static ArrayList<Faculty> facultyList;
-  public static HashMap<String, DayOfWeek> strToDayOfWeek = new HashMap<String, DayOfWeek>(){{
-    put("Sunday", DayOfWeek.SUNDAY);
-    put("Monday", DayOfWeek.MONDAY);
-    put("Tuesday", DayOfWeek.TUESDAY);
-    put("Wednesday", DayOfWeek.WEDNESDAY);
-    put("Thursday", DayOfWeek.THURSDAY);
-    put("Friday", DayOfWeek.FRIDAY);
-    put("Saturday", DayOfWeek.SATURDAY);
-  }};
-
-  public static HashMap<DayOfWeek, String> dayOfWeekToShortStr = new HashMap<DayOfWeek, String>(){{
-    put(DayOfWeek.SUNDAY, "SU");
-    put(DayOfWeek.MONDAY, "MO");
-    put(DayOfWeek.TUESDAY, "TU");
-    put(DayOfWeek.WEDNESDAY, "WE");
-    put(DayOfWeek.THURSDAY, "TH");
-    put(DayOfWeek.FRIDAY, "FR");
-    put(DayOfWeek.SATURDAY, "SA");
-  }};
-
-  public static HashMap<String, String> shortBuildingToAddress = new HashMap<String, String>(){{//incomplete address list for whole school
-    put("Howe", "Wesley J. Howe Center, 1 Castle Point Terrace, Hoboken, NJ 07030, USA");//McLean Hall, River St, Hoboken, NJ 07030, USA
-    put("McLean", "McLean Hall, River St, Hoboken, NJ 07030, USA");
-    put("Peirce", "Morton-Peirce-Kidde Complex, 607 River St, Hoboken, NJ 07030, USA");
-    put("Carnegie", "Carnegie Laboratory, Hoboken, NJ 07030, USA");
-    put("Burchard", "524 River St #713, Hoboken, NJ 07030");
-    put("Kidde", "607 River St, Hoboken, NJ 07030");
-    put("Gateway North", "601 Hudson St, Hoboken, NJ 07030");
-    put("Edwin A. Stevens", "Hoboken, NJ 07030");
-    put("Gateway South", "601 Hudson St, Hoboken, NJ 07030");
-    put("Morton", "Morton-Peirce-Kidde Complex, 607 River St, Hoboken, NJ 07030, USA");
-    put("Babbio", "525 River St, Hoboken, NJ 07030");
-    put("North Building", "North Building, 1 Castle Point Terrace, Hoboken, NJ 07030, USA");//McLean Hall, River St, Hoboken, NJ 07030, USA
-  }};
 
   public static void SavedScheduleToCalendar(ArrayList<Faculty> facultyList1, Sheet sheet) throws IOException, GeneralSecurityException, ParseException {
     // Load faculty list
@@ -297,8 +264,8 @@ private static VEvent setupAndMakeEvent(String eventInput, String eventName, Str
       String location;
       int meetingPatternsArrLength = meetingPatternsArr.length;
 
-      location = (meetingPatternsArr.length < 3 ? "" : shortBuildingToAddress.containsKey(meetingPatternsArr[meetingPatternsArrLength-1].split(" ")[0]) 
-                                                     ? shortBuildingToAddress.get(meetingPatternsArr[meetingPatternsArrLength-1].split(" ")[0]) 
+      location = (meetingPatternsArr.length < 3 ? "" : CalendarHelperFunctions.shortBuildingToAddress.containsKey(meetingPatternsArr[meetingPatternsArrLength-1].split(" ")[0]) 
+                                                     ? CalendarHelperFunctions.shortBuildingToAddress.get(meetingPatternsArr[meetingPatternsArrLength-1].split(" ")[0]) 
                                                      : meetingPatternsArr[meetingPatternsArrLength-1].split(" \\d")[0]);//Defaults to short address if long address is not in hashmap
       
       Faculty instructor = null;
@@ -327,8 +294,8 @@ private static VEvent setupAndMakeEvent(String eventInput, String eventName, Str
       System.out.println(classesStartReccuringDate);
       System.out.println(classesEndReccuringDate);
 
-      LocalDate startLocalDate = getNextDateOfDayFromDate(classesStartReccuringDate, getStartDay(daysMeetingStrArr, classesStartReccuringDate));
-      LocalDate endLocalDate = getNextDateOfDayFromDate(classesStartReccuringDate, getStartDay(daysMeetingStrArr, classesStartReccuringDate));
+      LocalDate startLocalDate = CalendarHelperFunctions.getNextDateOfDayFromDate(classesStartReccuringDate, CalendarHelperFunctions.getStartDay(daysMeetingStrArr, classesStartReccuringDate));
+      LocalDate endLocalDate = CalendarHelperFunctions.getNextDateOfDayFromDate(classesStartReccuringDate, CalendarHelperFunctions.getStartDay(daysMeetingStrArr, classesStartReccuringDate));
     
       //fix for when meetingPatternsArr contains Start/End date as well
       String[] startAndEndTimesInHalfs;
@@ -357,74 +324,11 @@ private static VEvent setupAndMakeEvent(String eventInput, String eventName, Str
       // System.out.println(endDateTimeStr);
       // //DateTime endDateTime = new DateTime(endDateTimeStr);
 
-      String recurrence = makeRecurrenceString(classesEndReccuringDate, daysMeetingStrArr);
+      String recurrence = CalendarHelperFunctions.makeRecurrenceString(classesEndReccuringDate, daysMeetingStrArr);
 
       
       
-      return makeEvent(eventName, location, description, recurrence, LocalDateTime.of(startLocalDate, startLocalTime), 
+      return CalendarHelperFunctions.makeEvent(eventName, location, description, recurrence, LocalDateTime.of(startLocalDate, startLocalTime), 
                                                                            LocalDateTime.of(endLocalDate, endLocalTime));
-  }
-  private static VEvent makeEvent(String eventName, String location, String description, String recurrence, LocalDateTime startDateTime, LocalDateTime endDateTime) {
-      UidGenerator ug = new RandomUidGenerator();
-      Uid uid = ug.generateUid();
-      
-      Recur recur = new Recur(recurrence, false);
-      RRule rrule = new RRule(recur);
-      VEvent vEvent = new VEvent(startDateTime, endDateTime, eventName).withProperty(uid)
-                                                                      .withProperty(new Description(description))
-                                                                      .withProperty(new Location(location))
-                                                                      .withProperty(rrule).getFluentTarget();
-      return vEvent;
-  }
-
-  public static DayOfWeek getStartDay(String[] daysMeetingStrArr, LocalDate classesStartReccuringDate) {
-    // Needs to get the start day by checking which day is nearest to the classesStartReccuringDate
-    // S M T W T F S
-
-    //Loop through the days in the daysMeetingStrArr to check if they are on the classesStartReccuringDate or after it
-    //if one of them is then pick that one
-    //if not pick the first day in the list
-
-    for (String s : daysMeetingStrArr) {
-      DayOfWeek meetingDay = strToDayOfWeek.get(s);
-      DayOfWeek reccuringStartDay = classesStartReccuringDate.getDayOfWeek();
-      if (reccuringStartDay.getValue() <= meetingDay.getValue()) {
-        return meetingDay;
-      }
-    }
-
-
-
-    return strToDayOfWeek.get(daysMeetingStrArr[0]);
-  }
-
-  public static LocalDate getNextDateOfDayFromDate(LocalDate startingDate, DayOfWeek dayOfWeek) {//Bug that disallows for classes to start on the starting day given
-    LocalDate result = startingDate;
-    int currentDay = result.getDayOfWeek().getValue();
-    int targetDay = dayOfWeek.getValue();
-    int daysToAdd = targetDay - currentDay; 
-    if (currentDay > targetDay) {
-      daysToAdd += 7;
-    }
-    result = result.plusDays(daysToAdd);
-    // while (result.getDayOfWeek() != dayOfWeek) {
-    //   result = result.plusDays(1);
-    // }
-    return result;
-  }
-
-  public static String makeRecurrenceString(LocalDate localEndDate, String[] daysMeeting) {
-
-    //RRULE:FREQ=WEEKLY;UNTIL=20230820T015615Z;WKST=SU;BYDAY=TU,TH //year month day
-    String result = "FREQ=WEEKLY;UNTIL=" + localEndDate.format(DateTimeFormatter.ofPattern("yyyyMMdd")) + "T000000Z;WKST=SU;BYDAY=";//SU
-    for (int i = 0; i < daysMeeting.length; i++) {
-      if (i == 0) {
-        result += dayOfWeekToShortStr.get(strToDayOfWeek.get(daysMeeting[i]));
-      } else {
-        result += "," + dayOfWeekToShortStr.get(strToDayOfWeek.get(daysMeeting[i]));
-      }
-    }
-    System.out.println(result);
-    return result;
   }
 }
